Add response_mate rake task to get response output
Why?
---

In order to minimize maintenance of the example responses.
Currently we're using constants in lib/resources (an idea borrowed by
github) to keep example response output.
When a response changes the constant has to be changed again.
So @ctrochalakis came up with the idea of some task to automate this.
The initial implementation is this.

Usage
-----
`rake response_mate:clear  #To remove old response files`
`rake response_mate:fetch #To grab the new response files`

* To add new requests tweak `requests.yml`
* To include a response in the markup `<%= prepared_response :shops %>` (in this
case we have named a response shops in `requests.yml`)